requirements_dev.txt
pytest==2.9.2
pytest-aiohttp==0.1.3
pytest-watch==4.1.0
